gateway/pyproject.toml
[project]
name = "gateway"
version = "0.1.0"
description = "Gateway"
authors = [
    {name = "thehecktour"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "facebook-business (>=22.0.5,<23.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "fastapi[standard] (>=0.115.12,<0.116.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
